Module Monoid.AddUnit

Augment a semigroup with a unit to form a monoid

Parameters

M : sig ... end

Signature

include S

include Putil.S

type t

val pp : Stdlib.Format.formatter -> t -> unit

val mul : t -> t -> t

val equal : t -> t -> bool

val unit : t

val lift : M.t -> t

Lift an element of the underlying semigroup into the monoid

val drop : t -> M.t option

Drop an element of the monoid into the underyling semigroup; evaluates to None on the unit.