Module Semilattice.Table

Hash table where values stored in the table belong to a semilattice. For each key k in the table ht, find ht k is the least upper bound of all values that have been added with key k.

Parameters

D : Stdlib.Hashtbl.HashedType

C : Sig.Semilattice.S

Signature

type t

val create : int -> t

val mem : t -> D.t -> bool

val clear : t -> unit

val find : t -> D.t -> C.t

val add : t -> D.t -> C.t -> unit