Module Apak__Lattice

Operations on lattices. A lattice is a set equipped with a partial order which has binary least upper bounds and greatest lower bounds.

module type S = Apak.Sig.Lattice.S

module Ordered : sig ... end

module LiftSubset : functor (S : Apak.Putil.Set.S) -> sig ... end

Powerset lattice under the subset ordering.

module Dual : functor (L : Apak.Sig.Lattice.S) -> Apak.Sig.Lattice.S with type t = L.t

Flip the order in a lattice

module FunctionSpace : sig ... end

module Bounded : sig ... end

A bounded lattice has a least and greatest element