Module Apak__Lattice.Bounded

A bounded lattice has a least and greatest element

type 'a bounded =

| Top

| Bottom

| Value of 'a

module Lift : functor (L : Apak.Sig.Lattice.S) -> Apak.Sig.Lattice.Bounded.S with type t = L.t bounded

Augment a lattice with least and greatest elements

module Ordered : sig ... end

module LiftSubset : functor (S : Apak.Putil.Set.S) -> sig ... end

Boolean algebra of finite + cofinite sets. The type S.elt is assumed to have infinitely many inhabitants (so finite sets and cofinite sets are disjoint).

module FunctionSpace : sig ... end