Module FunctionSpace.Total

Total function spaces where all but finitely many elements map to the same "default" value.

module type S = sig ... end

module LiftMap : functor (M : Apak.Putil.Map.S) -> functor (Codomain : Apak.Sig.Lattice.Bounded.S) -> S with type dom = M.key and type cod = Codomain.t

LiftMap(M)(Codomain) constructs a total function space lattice where M is used as the underlying map representation. This is useful (for example) with tagged types, since Patricia trees support fast merging.

module Make : functor (Domain : Apak.Putil.Ordered) -> functor (Codomain : Apak.Sig.Lattice.Bounded.S) -> S with type dom = Domain.t and type cod = Codomain.t

module Smash : functor (Domain : Apak.Putil.Ordered) -> functor (Codomain : Apak.Sig.Lattice.Bounded.S) -> S with type dom = Domain.t and type cod = Codomain.t

Total function space where every function that maps at least one point to bottom is identified with bottom

module Ordered : sig ... end