Module Total.LiftMap

LiftMap(M)(Codomain) constructs a total function space monoid where M is used as the underlying map representation. This is useful (for example) with tagged types, since Patricia trees support fast merging.

Parameters

M : Apak.Putil.Map.S

Codomain : Apak.Sig.Monoid.S

Signature

merge m f g constructs the function which maps each x to m (f x) (g x)

include S

include Apak.Sig.Monoid.S

include Apak.Putil.S

type t

val pp : Stdlib.Format.formatter -> t -> unit

val mul : t -> t -> t

val equal : t -> t -> bool

val unit : t

type dom

type cod

val eval : t -> dom -> cod

val map : (cod -> cod) -> t -> t

val update : dom -> cod -> t -> t

val enum : t -> (dom * cod) BatEnum.t

val support : t -> dom BatEnum.t

val merge : (cod -> cod -> cod) -> t -> t -> t

merge m f g constructs the function which maps each x to m (f x) (g x)

val default : t -> cod

default f is the value to which f maps all elements of the domain which are not in the support of f.

val const : cod -> t

const k constructs a function where every element of the domain is mapped to k