Module Apak__Putil

Apak pervasives

module type S = sig ... end

module type OrderedMix = sig ... end

module type Ordered = sig ... end

module Set : sig ... end

module Map : sig ... end

module MonoMap : sig ... end

Monomorphic maps

module TotalFunction : sig ... end

Total function spaces where all but finitely many elements map to the same "default" value.

module Hashed : sig ... end

module type CoreTypeBasis = sig ... end

module type CoreType = sig ... end

module MakeCoreType : functor (T : CoreTypeBasis) -> CoreType with type t = T.t

module PString : CoreType with type t = string

module PInt : CoreType with type t = int

module PChar : CoreType with type t = char

module PUnit : CoreType with type t = unit

val set_load_path : string -> unit

val set_temp_dir : string -> unit

val find_file : string -> string option

Search for a file named file in all the directories in the load path. Raise File_not_found if no such file exists.

val with_temp_filename : string -> string -> (string -> 'a) -> 'a

val with_temp_file : string -> string -> (string -> Stdlib.out_channel -> 'a) -> 'a

val with_temp_dir : string -> (string -> 'a) -> 'a

val compare_tuple : int Stdlib.Lazy.t list -> int