Module Apak__Semilattice

Operations on semilattices. A semilattice is a set equipped with an associative, commutative, idempotent binary operation

module type S = Apak.Sig.Semilattice.S

module Ordered : sig ... end

module FunctionSpace : sig ... end

module Table : functor (D : Stdlib.Hashtbl.HashedType) -> functor (C : Apak.Sig.Semilattice.S) -> sig ... end

Hash table where values stored in the table belong to a semilattice. For each key k in the table ht, find ht k is the least upper bound of all values that have been added with key k.

module Bounded : sig ... end

Bounded semilattices are semilattices with a least element, bottom