Module Apak__Semilattice.Bounded

Bounded semilattices are semilattices with a least element, bottom

type 'a bounded =

| Bottom

| Value of 'a

val pp_bounded : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a bounded -> unit

module Lift : functor (S : Apak.Sig.Semilattice.S) -> Apak.Sig.Semilattice.Bounded.S with type t = S.t bounded

Augment a semilattice with a least element

module Monoid : functor (S : Apak.Sig.Semilattice.Bounded.S) -> Apak.Sig.Monoid.S with type t = S.t

A bounded semilattice forms a monoid, taking multiplication to be join and the unit to be bottom.

module Ordered : sig ... end