Module Bounded.Monoid

A bounded semilattice forms a monoid, taking multiplication to be join and the unit to be bottom.

Parameters

S : Apak.Sig.Semilattice.Bounded.S

Signature

include Apak.Putil.S

type t

val pp : Stdlib.Format.formatter -> t -> unit

val mul : t -> t -> t

val equal : t -> t -> bool

val unit : t